+++ Makefile
@@ -58,11 +58,5 @@ install: all openlibm.pc
 	mkdir -p $(DESTDIR)$(includedir)/openlibm
 	cp -f -a libopenlibm.$(SHLIB_EXT)* $(DESTDIR)$(shlibdir)/
 	cp -f -a libopenlibm.a $(DESTDIR)$(libdir)/
-	cp -f -a src/openlibm.h $(DESTDIR)$(includedir)/
+	cp -f -a include/openlibm*.h $(DESTDIR)$(includedir)/
 	cp -f -a openlibm.pc $(DESTDIR)$(libdir)/pkgconfig/
-ifneq ($(wildcard $(ARCH)/bsd_asm.h),)
-	cp -f -a $(ARCH)/bsd_asm.h $(DESTDIR)$(includedir)/openlibm/
-endif
-ifneq ($(wildcard $(ARCH)/bsd_cdefs.h),)
-	cp -f -a $(ARCH)/bsd_cdefs.h $(DESTDIR)$(includedir)/openlibm/
-endif
