+++ i387/bsd_ieeefp.h
@@ -1,2 +1,2 @@
-#include "include/cdefs-compat.h"
+#include "cdefs-compat.h"
 #include "amd64/bsd_ieeefp.h"
