+++ src/cdefs-compat.h
